Use skyCatalog config to initialize Gaia stuff. Move GaiaObject and GaiaCollection from test program to gaia_object.py under objects directory
