+++ python/desc/skycatalogs/objects/__init__.py
@@ -1,1 +1,2 @@
 from .base_object import *
+from .gaia_object import *
